requirements.txt
requests>=2.31.0
loguru>=0.7.0
eth-account>=0.8.0
hyperliquid-python-sdk>=0.2.0
python-dotenv>=1.0.0
